Enable emacs socket activation
This should prevent high CPU load on my non-NixOS servers, which try
to invoke non-Nix emacs on my Emacs configuration, which takes long to
native-compile and due to that the systemd service timeouts, is
automatically restarted and the compilation starts again. This
continues indefinitely.

By providing a service with Nix Emacs, if the same happens, I can
either fix it by modifying the service config in Nix or by running the
emacs manually on the server. Note that running manually the non-Nix
emacs requires running it with its full path, i.e. /usr/bin/emacs,
because the Emacs from nix overrides the non-Nix one. Therefore, it is
easy to not notice the problem for long time.